Create README.md
Welcome to the wonderful world of Pokémon! We want to keep the world wonderful, but due to climate change and the impact human beings have had on habitat loss we've had to make changes. Our Pokémon world region has decided that to accomplish this, Pokémon trainers will only be allowed to keep one type of each Pokémon.

The Pokémon Box technology needs upgrading, the code provided does not limit the addition of Pokémon with the same name.
Thats why, we created a custom exception class, called PokemonAlreadyExistsException that can be thrown when the add method in PokemonBox. The constructors also need upgrading to throw an exception, rather than shut down the program, so we fixed it. We used  IllegalArgumentException in the Pokemon class constructors.